# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  member: "member",
} as const;

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface SignupBody {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role?: UserRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface Project {
  id: number;
  name: string;
  description: string;
  createdAt: string;
  ownerId: number;
  memberIds: number[];
  members: User[];
  taskCount: number;
  completedCount: number;
}

export interface CreateProjectBody {
  /** @minLength 1 */
  name: string;
  description: string;
  memberIds: number[];
}

export interface UpdateProjectBody {
  name?: string;
  description?: string;
  memberIds?: number[];
}

export interface Task {
  id: number;
  projectId: number;
  projectName: string;
  title: string;
  description: string;
  status: TaskStatus;
  dueDate?: string | null;
  assigneeId?: number | null;
  assignee?: User | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTaskBody {
  /** @minLength 1 */
  title: string;
  description: string;
  status?: TaskStatus;
  dueDate?: string | null;
  assigneeId?: number | null;
}

export interface UpdateTaskBody {
  title?: string;
  description?: string;
  status?: TaskStatus;
  dueDate?: string | null;
  assigneeId?: number | null;
}

export interface DashboardSummary {
  total: number;
  pending: number;
  inProgress: number;
  completed: number;
  overdue: number;
  dueSoon: number;
  projectCount: number;
}