# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Team Task Manager API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: users
    description: User directory
  - name: projects
    description: Project management (admin)
  - name: tasks
    description: Task management
  - name: dashboard
    description: Dashboard summaries

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "200":
          description: Signed up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (for assignment)
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects visible to the current user
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project (admin only)
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (admin only)
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /projects/{projectId}/tasks:
    get:
      operationId: listProjectTasks
      tags: [tasks]
      summary: List all tasks for a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task in a project (admin only)
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task. Members may update only status on their own tasks.
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task (admin only)
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /tasks/me:
    get:
      operationId: listMyTasks
      tags: [tasks]
      summary: List tasks assigned to the current user
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Status summary for the current user
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/recent:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Recent task activity visible to the user
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    UserRole:
      type: string
      enum: [admin, member]

    TaskStatus:
      type: string
      enum: [pending, in_progress, completed]

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
      required: [id, name, email, role]

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]

    SignupBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          $ref: "#/components/schemas/UserRole"
      required: [name, email, password]

    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]

    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        ownerId:
          type: integer
        memberIds:
          type: array
          items:
            type: integer
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
        taskCount:
          type: integer
        completedCount:
          type: integer
      required: [id, name, description, createdAt, ownerId, memberIds, members, taskCount, completedCount]

    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        memberIds:
          type: array
          items:
            type: integer
      required: [name, description, memberIds]

    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        memberIds:
          type: array
          items:
            type: integer

    Task:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        dueDate:
          type: string
          nullable: true
          format: date-time
        assigneeId:
          type: integer
          nullable: true
        assignee:
          allOf:
            - $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, projectId, projectName, title, description, status, createdAt, updatedAt]

    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        dueDate:
          type: string
          nullable: true
          format: date-time
        assigneeId:
          type: integer
          nullable: true
      required: [title, description]

    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        dueDate:
          type: string
          nullable: true
          format: date-time
        assigneeId:
          type: integer
          nullable: true

    DashboardSummary:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        inProgress:
          type: integer
        completed:
          type: integer
        overdue:
          type: integer
        dueSoon:
          type: integer
        projectCount:
          type: integer
      required: [total, pending, inProgress, completed, overdue, dueSoon, projectCount]