# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatus } from "./taskStatus";
import type { User } from "./user";

export interface Task {
  id: number;
  projectId: number;
  projectName: string;
  title: string;
  description: string;
  status: TaskStatus;
  dueDate?: Date | null;
  assigneeId?: number | null;
  assignee?: User | null;
  createdAt: Date;
  updatedAt: Date;
}