# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatus } from "./taskStatus";

export interface UpdateTaskBody {
  title?: string;
  description?: string;
  status?: TaskStatus;
  dueDate?: Date | null;
  assigneeId?: number | null;
}